pyproject.toml
[tool.poetry]
name = "rag_demo"
version = "0.1.0"
description = ""
authors = [
  "Ernő Erdélyi <erno@componentsoft.io>",
  "Tamás Kiss <tamas.kiss@componentsoft.io>",
  "Kristóf Osztopáni <kristof@componentsoft.io>"
]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chromadb = "^0.4.22"
langchain = "^0.1.7"
langchain-community = "^0.0.20"
langchain-core = "^0.1.23"
langchain-mistralai = "^0.0.4"
langchain-openai = "^0.0.6"
langchainhub = "^0.1.14"
langfuse = "^2.20.1"
numpy = "^1.26.2"
pypdf = "^4.1.0"
python-dotenv = "^0.21.1"
sentence-transformers = "^2.5.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^6.1.0"
mypy = "^1.8.0"
flake8-pyproject = "^1.2.3"
pre-commit = "^3.6.0"
types-python-dateutil = "^2.8.19.14"

[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = 0
line_length = 88

[tool.black]
line-length = 88
target-version = ['py310']

[tool.flake8]
max-line-length = 88
exclude = [
    ".git",
    "__pycache__",
    "chroma",
    "data"
]
extend-ignore = []

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
show_error_codes = true
python_version = "3.10"
exclude = [
    ".git",
    "__pycache__"
]
